Reorganize templates in KAppTemplate in order to provide an easier to understand taxonomy.
I have changed the templates category path into Technology/Type/Language instead of Language/Type+Technology as it was right now. I'd like to delay the language choice until the end, in an attempt to remove a little the initial obligation to select/marry the language or technology.

CCMAIL: annma@kde.org

svn path=/trunk/KDE/kdesdk/kapptemplate/; revision=1248732