Only replace keywords in text files
Non-text files such as PNG and compressed SVG were being read in for
keyword substitution and then written back out as text, even if no
substitions were made. As a result they were corrupt in the generated
project.

If the mimetype of a file does not descend from text/plain, the file
will no longer be considered for keyword substitution.

Addtionally, only changed files are written.

BUG: 502500